djangoapp/requirements.txt
asgiref==3.8.1
bcc==0.1.10
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
traitlets==5.14.3
traittypes==0.2.1
